correct foldsHistogram computation ref:31c3af5ca33270e6b6b0db2969a55f76b33ab9c4

+++ catboost/cuda/methods/kernel/pointwise_hist2_one_byte_8bit.cu
@@ -94,8 +94,8 @@ namespace NKernel
 
             #pragma unroll
             for (int i = 0; i < 4; i++) {
-                const int f = ((2 * i + threadIdx.x ) & 6);
-                const int bin = bfe(ci, 24 - (f << 2), 8);
+                const int f = ((2 * i + threadIdx.x) & 6);
+                const uchar bin = bfe(ci, 24 - (f << 2), 8);
 
                 if (bin != mostRecentBin[i]) {
                     int offset = f;
